requirements.txt
tweepy
scapy
pysocks
dnslib
pycrypto